Port chrono test to new integration test style
