+++ schemars/Cargo.toml
@@ -40,8 +40,9 @@ serde = { version = "1.0", features = ["derive"] }
 jsonschema = { version = "0.18.1", default-features = false }
 snapbox = { version = "0.6.17", features = ["json"] }
 
-arrayvec07 = { version = "0.7", features = ["serde"], package = "arrayvec"}
-bytes1 = { version = "1.0", features = ["serde"], package = "bytes" }
+arrayvec07 = { version = "0.7", default-features = false, features = ["serde"], package = "arrayvec"}
+bytes1 = { version = "1.0", default-features = false, features = ["serde"], package = "bytes" }
+chrono04 = { version = "0.4", default-features = false, features = ["serde"], package = "chrono" }
 
 [features]
 default = ["derive", "std"]
@@ -75,10 +76,6 @@ required-features = ["std"]
 [[test]]
 name = "ui"
 required-features = ["_ui_test"]
-
-[[test]]
-name = "chrono"
-required-features = ["chrono04"]
 
 [[test]]
 name = "indexmap"
