--- schemars/tests/chrono.rs
@@ -1,19 +1,0 @@
-mod util;
-use chrono04::prelude::*;
-use schemars::JsonSchema;
-use util::*;
-
-#[allow(dead_code)]
-#[derive(JsonSchema)]
-struct ChronoTypes {
-    weekday: Weekday,
-    date_time: DateTime<Utc>,
-    naive_date: NaiveDate,
-    naive_date_time: NaiveDateTime,
-    naive_time: NaiveTime,
-}
-
-#[test]
-fn chrono_types() -> TestResult {
-    test_default_generated_schema::<ChronoTypes>("chrono-types")
-}
