+++ schemars/tests/integration/snapshots/schemars/tests/integration/chrono.rs~chrono.de.json
@@ -29,7 +29,7 @@
     },
     "naive_time": {
       "type": "string",
-      "format": "partial-date-time"
+      "format": "partial-time"
     }
   },
   "required": [
