+++ schemars/tests/integration/snapshots/schemars/tests/integration/chrono.rs~chrono.ser.json
@@ -1,0 +1,42 @@
+{
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "title": "ChronoTypes",
+  "type": "object",
+  "properties": {
+    "weekday": {
+      "type": "string",
+      "enum": [
+        "Mon",
+        "Tue",
+        "Wed",
+        "Thu",
+        "Fri",
+        "Sat",
+        "Sun"
+      ]
+    },
+    "date_time": {
+      "type": "string",
+      "format": "date-time"
+    },
+    "naive_date": {
+      "type": "string",
+      "format": "date"
+    },
+    "naive_date_time": {
+      "type": "string",
+      "format": "partial-date-time"
+    },
+    "naive_time": {
+      "type": "string",
+      "format": "partial-time"
+    }
+  },
+  "required": [
+    "weekday",
+    "date_time",
+    "naive_date",
+    "naive_date_time",
+    "naive_time"
+  ]
+}
